Initial Commit of PygTailz
This initial commit is part of me transferring my current, and some older, projects to Github. PygTailz was originally an attempt to create a library that made Pygame easier to use by "setting up" things like user input, sprite placement & movement, a state managers, etc. for the user in the form of Python classes.